requirements.txt
# Core Python dependencies
google-genai>=1.33.0
python-dotenv>=1.0.0

# LangChain ecosystem - latest versions
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# Vector store and embeddings
langchain-chroma>=0.1.0
chromadb>=0.5.0
langchain-nomic>=0.1.0
nomic>=3.0.0

# Data processing and utilities
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Web interface - Required
streamlit>=1.28.0
streamlit-chat>=0.1.0
plotly>=5.15.0
streamlit-plotly-events>=0.0.6

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Utility packages
tqdm>=4.65.0
pathlib2>=2.3.0
typing-extensions>=4.5.0
pydantic>=2.0.0

# For enhanced processing
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Advanced features
jupyter>=1.0.0
ipywidgets>=8.0.0